refactor

+++ api/app/Http/Controllers/CommentController.php
@@ -2,61 +2,30 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\StoreCommentRequest;
-use App\Http\Requests\UpdateCommentRequest;
-use App\Http\Requests\DestroyCommentRequest;
+use App\Http\Requests\SaveCommentRequest;
 use App\Models\Comment;
-use App\Models\Post;
+use App\Http\Resources\CommentResource;
 
 class CommentController extends Controller
 {
-    public function index(Post $post)
-    {
-        $comments = $post->comments;
-
-        return [
-            'data' => [
-                'comments' => $comments,
-            ],
-        ];
-    }
-
-    public function store(StoreCommentRequest $request, Post $post)
-    {
-        $comment = new Comment;
-
-        $comment->content = $request->content;
-
-        $comment->user()->associate($request->user());
-        $comment->post()->associate($post);
-
-        $comment->save();
-
-        return [
-            'data' => [
-                'id' => $comment->id,
-            ]
-        ];
-    }
-
     public function show(Comment $comment)
     {
-        return [
-            'data' => [
-                'comment' => $comment,
-            ],
-        ];
+        return new CommentResource($comment);
     }
 
-    public function update(UpdateCommentRequest $request, Comment $comment)
+    public function update(SaveCommentRequest $request, Comment $comment)
     {
+        $this->authorize('update', $comment);
+
         $comment->update($request->only(['content']));
 
         return response()->noContent();
     }
 
-    public function destroy(DestroyCommentRequest $request, Comment $comment)
+    public function destroy(Comment $comment)
     {
+        $this->authorize('delete', $comment);
+
         $comment->delete();
 
         return response()->noContent();
